requirements.txt
numpy>=1.26
pandas>=2.2
torch>=2.2
transformers>=4.43
pydantic>=2.8
pyyaml>=6.0
tqdm>=4.66
scipy>=1.12
python-dotenv
openai
